fix mission repo

+++ as-15-8/as-15-8-pod/src/main/java/repositories/MissionRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MissionRepository extends JpaRepository<Mission, Integer>, MissionRepositoryCustom{
+public interface MissionRepository extends JpaRepository<Mission, Long>, MissionRepositoryCustom{
 
     List<Mission> findByOrigIdIn(List<String> ids);
